Restructure everything and add more comments
