+++ src/pages/api/examples.ts
@@ -1,6 +1,6 @@
 import { type NextApiRequest, type NextApiResponse } from "next";
 
-import { prisma } from "../../server/db/client";
+import { prisma } from "../../server/db";
 
 const examples = async (req: NextApiRequest, res: NextApiResponse) => {
   const examples = await prisma.example.findMany();
