+++ src/pages/api/trpc/[trpc].ts
@@ -1,13 +1,13 @@
 import { createNextApiHandler } from "@trpc/server/adapters/next";
 
 import { env } from "../../../env/server.mjs";
-import { createContext } from "../../../server/api/trpc/context";
+import { createInternalTrpcContext } from "../../../server/api/trpc";
 import { appRouter } from "../../../server/api/root";
 
 // export API handler
 export default createNextApiHandler({
   router: appRouter,
-  createContext,
+  createContext: createInternalTrpcContext,
   onError:
     env.NODE_ENV === "development"
       ? ({ path, error }) => {
