+++ src/server/db.ts
@@ -1,6 +1,6 @@
 import { PrismaClient } from "@prisma/client";
 
-import { env } from "../../env/server.mjs";
+import { env } from "../env/server.mjs";
 
 declare global {
   // eslint-disable-next-line no-var
